requirements.txt
chardet==5.2.0
pdfminer2
Flask==3.1.1